Added functionalities for language toggle
Makes it so that chosing a language fx "en" adds /en to the url and invokes it new url. Which means that it will show the localized page kundeservice.dk/en which has the translated strings connected to it. The state is saved in localstorage and gets the state from there to workaround the page reloads.